move temp examples fixture to be in main conftest, and enable repo-wide access
