+++ h2integrate/converters/solar/test/conftest.py
@@ -2,7 +2,11 @@ import pytest
 
 from h2integrate import EXAMPLE_DIR
 
-from test.conftest import temp_dir, pytest_collection_modifyitems  # noqa: F401
+from test.conftest import (  # noqa: F401
+    temp_dir,
+    temp_copy_of_example,
+    pytest_collection_modifyitems,
+)
 
 
 @pytest.fixture
