+++ h2integrate/converters/test/conftest.py
@@ -4,7 +4,11 @@ Pytest configuration file.
 
 import os
 
-from test.conftest import temp_dir, pytest_collection_modifyitems  # noqa: F401
+from test.conftest import (  # noqa: F401
+    temp_dir,
+    temp_copy_of_example,
+    pytest_collection_modifyitems,
+)
 
 
 def pytest_sessionstart(session):
