+++ h2integrate/resource/utilities/test/conftest.py
@@ -2,7 +2,11 @@ import os
 
 from hopp import TEST_ENV_VAR
 
-from test.conftest import temp_dir, pytest_collection_modifyitems  # noqa: F401
+from test.conftest import (  # noqa: F401
+    temp_dir,
+    temp_copy_of_example,
+    pytest_collection_modifyitems,
+)
 
 
 def pytest_sessionstart(session):
