+++ h2integrate/resource/wind/test/conftest.py
@@ -5,4 +5,8 @@ from h2integrate.resource.test.conftest import (  # noqa: F401
     pytest_sessionfinish,
 )
 
-from test.conftest import temp_dir, pytest_collection_modifyitems  # noqa: F401
+from test.conftest import (  # noqa: F401
+    temp_dir,
+    temp_copy_of_example,
+    pytest_collection_modifyitems,
+)
