+++ h2integrate/storage/battery/test/conftest.py
@@ -1,3 +1,7 @@
 from h2integrate.storage.test.conftest import plant_config  # noqa: F401
 
-from test.conftest import temp_dir, pytest_collection_modifyitems  # noqa: F401
+from test.conftest import (  # noqa: F401
+    temp_dir,
+    temp_copy_of_example,
+    pytest_collection_modifyitems,
+)
