+++ h2integrate/storage/hydrogen/test/conftest.py
@@ -6,7 +6,11 @@ import os
 
 from h2integrate.storage.test.conftest import plant_config  # noqa: F401
 
-from test.conftest import temp_dir, pytest_collection_modifyitems  # noqa: F401
+from test.conftest import (  # noqa: F401
+    temp_dir,
+    temp_copy_of_example,
+    pytest_collection_modifyitems,
+)
 
 
 def pytest_sessionstart(session):
